+++ src/CloudMigrator.Cli/Program.cs
@@ -10,6 +10,7 @@ Console.CancelKeyPress += (_, e) =>
 
 var rootCmd = new RootCommand("CloudMigrator - OneDrive から SharePoint へのファイル移行ツール");
 rootCmd.Add(TransferCommand.Build());
+rootCmd.Add(TransferStatusCommand.Build());
 rootCmd.Add(RebuildSkipListCommand.Build());
 rootCmd.Add(WatchdogCommand.Build());
 rootCmd.Add(QualityMetricsCommand.Build());